+++ app/models/omni_event/log.rb
@@ -5,14 +5,16 @@ module OmniEvent
     belongs_to :loggable, polymorphic: true, optional: true
     has_one_attached :payload_debug
 
+    serialize :metadata, coder: JSON
+
     after_create_commit :dispatch_external_monitoring
 
     private
 
     def dispatch_external_monitoring
-      OmniEvent::NewRelicJob.perform_later(self.attributes)
+      OmniEvent::NewRelicJob.perform_later(attributes)
     rescue => e
-      Rails.logger.error "[OmniEvent] Failed to enqueue job: #{e.message}"
+      Rails.logger.error "[OmniEvent] Failed to enqueue NewRelicJob: #{e.message}"
     end
   end
 end