+++ lib/omni_event.rb
@@ -2,6 +2,8 @@ require "omni_event/version"
 require "omni_event/engine"
 require "omni_event/configuration"
 require "omni_event/base_processor"
+require "omni_event/process_dispatcher"
+require "omni_event/signature_verifier"
 
 module OmniEvent
   class << self