+++ spec/dummy/config/environment.rb
@@ -1,0 +1,4 @@
+ENV['RAILS_ENV'] ||= 'test'
+
+require_relative 'application'
+Rails.application.initialize!